Fix conflict

+++ src/restful/controllers/adminController.js
@@ -133,17 +133,18 @@ class AdminController {
     const { user_id } = req.params;
     try {
       const db = admin.firestore();
-      const userDoc = await db.collection("users").doc(user_id).get();
-      const userData = userDoc.data();
-      const imageUrl = userData?.imageUrl;
       await admin.auth().deleteUser(user_id);
-      await db.collection("users").doc(user_id).delete();
-      if (imageUrl) {
-        const fileName = imageUrl.split("/").pop();
-        const bucket = admin.storage().bucket();
-        await bucket.file(`profile/picture/${fileName}`).delete();
+      const userDoc = await db.collection("users").doc(user_id).get();
+      if(userDoc){
+        const userData = userDoc.data();
+        const imageUrl = userData?.imageUrl;
+        await db.collection("users").doc(user_id).delete();
+        if (imageUrl) {
+          const fileName = imageUrl.split("/").pop();
+          const bucket = admin.storage().bucket();
+          await bucket.file(`profile/picture/${fileName}`).delete();
+        }
       }
-
       util.statusCode = 200;
       util.setSuccess(200, "User deleted successfully!");
       return util.send(res);
@@ -154,6 +155,64 @@ class AdminController {
       return util.send(res);
     }
   }
+
+  static async getAllOpportunities(req, res) {
+    const db = admin.firestore();
+  
+    try {
+      const applicationsSnapshot = await db.collectionGroup("applications").get();
+      const opportunityApplicationsCount = {};
+      
+      // Count the number of applications for each opportunity
+      applicationsSnapshot.forEach((doc) => {
+        const applicationData = doc.data();
+        const opportunityId = applicationData.opportunity_id;
+        if (opportunityApplicationsCount[opportunityId]) {
+          opportunityApplicationsCount[opportunityId]++;
+        } else {
+          opportunityApplicationsCount[opportunityId] = 1;
+        }
+      });
+  
+      // Fetch opportunities separately
+      const opportunitiesSnapshot = await db.collection("opportunities").where("status", "!=", "archived").get();
+  
+      // Fetch user data for each opportunity asynchronously
+      const opportunitiesDataPromises = opportunitiesSnapshot.docs.map(async (doc) => {
+        const opportunityData = doc.data();
+        const userId = opportunityData.user_id;
+        const userRef = db.collection("users").doc(userId);
+        const userDoc = await userRef.get();
+        const userData = userDoc.data();
+        const opportunityId = doc.id;
+  
+        return {
+          ...opportunityData,
+          numberOfApplications: opportunityApplicationsCount[opportunityId] || 0,
+          userName: userData ? userData.first_name + " " + userData.first_name : "Unknown",
+        };
+      });
+  
+      // Wait for all promises to resolve
+      const opportunitiesData1 = await Promise.all(opportunitiesDataPromises);
+  
+      if (opportunitiesData1.length > 0) {
+        util.statusCode = 200;
+        util.message = opportunitiesData1;
+        return util.send(res);
+      } else {
+        util.statusCode = 404;
+        util.message = "Not found";
+        return util.send(res);
+      }
+    } catch (error) {
+      console.error("Error fetching opportunities:", error);
+      util.statusCode = 500;
+      util.message = error.message || "Server error";
+      return util.send(res);
+    }
+  }
+  
 }
 
 exports.AdminController = AdminController;
